Moved plotting functions to plot subcommand
